refactor: make changes to use the new UserApi in the application
Change from the old server module usage of UserApi to use the new Kitsu module
