+++ app/src/main/java/com/chesire/nekome/injection/KitsuModule.kt
@@ -3,11 +3,9 @@ package com.chesire.nekome.injection
 import com.chesire.nekome.kitsu.api.auth.KitsuAuth
 import com.chesire.nekome.kitsu.api.library.KitsuLibrary
 import com.chesire.nekome.kitsu.api.search.KitsuSearch
-import com.chesire.nekome.kitsu.api.user.KitsuUser
 import com.chesire.nekome.server.api.AuthApi
 import com.chesire.nekome.server.api.LibraryApi
 import com.chesire.nekome.server.api.SearchApi
-import com.chesire.nekome.server.api.UserApi
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -36,10 +34,4 @@ abstract class KitsuModule {
      */
     @Binds
     abstract fun providesSearchApi(kitsuSearch: KitsuSearch): SearchApi
-
-    /**
-     * Binds the instance of [KitsuUser] to [UserApi].
-     */
-    @Binds
-    abstract fun providesUserApi(kitsuUser: KitsuUser): UserApi
 }
