+++ libraries/account/src/main/java/com/chesire/nekome/account/UserRepository.kt
@@ -4,13 +4,15 @@ import com.chesire.nekome.core.flags.Service
 import com.chesire.nekome.core.models.UserModel
 import com.chesire.nekome.database.dao.UserDao
 import com.chesire.nekome.core.Resource
-import com.chesire.nekome.server.api.UserApi
+import com.chesire.nekome.user.api.UserApi
+import com.chesire.nekome.user.api.UserEntity
 import timber.log.Timber
+import javax.inject.Inject
 
 /**
  * Repository to interact with user data.
  */
-class UserRepository(
+class UserRepository @Inject constructor(
     private val userDao: UserDao,
     private val userApi: UserApi
 ) {
@@ -24,16 +26,31 @@ class UserRepository(
      */
     suspend fun refreshUser(): Resource<UserModel> {
         val response = userApi.getUserDetails()
-        when (response) {
-            is Resource.Success -> userDao.insert(response.data)
-            is Resource.Error -> Timber.e("Error refreshing user")
+        return when (response) {
+            is Resource.Success -> {
+                val model = response.data.toUserModel()
+                userDao.insert(model)
+                Resource.Success(model)
+            }
+            is Resource.Error -> {
+                Timber.e("Error refreshing user")
+                Resource.Error(response.msg)
+            }
         }
-
-        return response
     }
 
     /**
      * Retrieves the user id from the database.
      */
     suspend fun retrieveUserId() = userDao.retrieveUserId(Service.Kitsu)
+
+    private fun UserEntity.toUserModel(): UserModel {
+        return UserModel(
+            userId,
+            name,
+            avatar,
+            coverImage,
+            service
+        )
+    }
 }
