+++ libraries/kitsu/user/src/main/java/com/chesire/nekome/kitsu/user/KitsuUser.kt
@@ -1,4 +1,4 @@
-package com.chesire.nekome.user
+package com.chesire.nekome.kitsu.user
 
 import com.chesire.nekome.core.Resource
 import com.chesire.nekome.kitsu.asError
@@ -7,8 +7,9 @@ import com.chesire.nekome.user.api.UserApi
 import com.chesire.nekome.user.api.UserEntity
 import com.chesire.nekome.user.api.UserEntityMapper
 import retrofit2.Response
+import javax.inject.Inject
 
-class KitsuUser(
+class KitsuUser @Inject constructor(
     private val userService: KitsuUserService,
     private val mapper: UserEntityMapper<KitsuUserEntity>
 ): UserApi {
