+++ libraries/kitsu/user/src/main/java/com/chesire/nekome/kitsu/user/KitsuUserEntityMapper.kt
@@ -1,12 +1,14 @@
-package com.chesire.nekome.user
+package com.chesire.nekome.kitsu.user
 
 import com.chesire.nekome.core.flags.RatingSystem
 import com.chesire.nekome.core.flags.Service
 import com.chesire.nekome.core.models.ImageModel
+import com.chesire.nekome.kitsu.user.KitsuUserEntity
 import com.chesire.nekome.user.api.UserEntity
 import com.chesire.nekome.user.api.UserEntityMapper
+import javax.inject.Inject
 
-class KitsuUserEntityMapper : UserEntityMapper<KitsuUserEntity> {
+class KitsuUserEntityMapper @Inject constructor() : UserEntityMapper<KitsuUserEntity> {
 
     override fun mapFromUserEntity(entity: UserEntity): KitsuUserEntity {
         return KitsuUserEntity(
